将所有Short/short类型改为Integer/int，便于使用。 只在序列化和反序列化的时候强制转换为short

+++ journalkeeper-core/src/main/java/com/jd/journalkeeper/core/entry/reserved/ScalePartitionsEntrySerializer.java
@@ -39,7 +39,7 @@ public class ScalePartitionsEntrySerializer implements Serializer<ScalePartition
         ByteBuffer buffer = ByteBuffer.wrap(bytes, Byte.BYTES, bytes.length - Byte.BYTES);
         List<Integer> partitionList = new LinkedList<>();
         while (buffer.hasRemaining()) {
-            partitionList.add(buffer.getInt());
+            partitionList.add(Short.valueOf(buffer.getShort()).intValue());
         }
         return new ScalePartitionsEntry(partitionList.stream().mapToInt(Integer::intValue).toArray());
     }
